--- Assets/Scenes/Example.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f1a4578fa7a844903a6ab02e9792748b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
