--- Assets/Scenes/Example/1.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d886ef3278dd947b8abc724b30df8204
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
