+++ Assets/Scripts/Menu/LobbyRoomUI.cs
@@ -16,6 +16,11 @@ public class LobbyRoomUI : MonoBehaviourPun {
     public GameObject readyPlayerItemPrefab;
     public GameObject unreadyPlayerItemPrefab;
 
+
+    //RoomList
+    public GameObject roomNamePrefab;
+    public Transform gridLayout;
+
     private Hashtable props;
 
     void Start() {
@@ -60,5 +65,32 @@ public class LobbyRoomUI : MonoBehaviourPun {
       } else {
         NetworkManager.instance.SetLocalPlayerProperty("Ready", true);
       }
+    }
+
+
+    //RoomList
+    public override void OnRoomListUpdate(List<RoomInfo> roomList)
+    {
+        for (int i = 0; i < gridLayout.childCount; i++)
+        {
+            if (gridLayout.GetChild(i).gameObject.GetComponentInChildren<Text>().text == roomList[i].Name)
+            {
+                Destroy(gridLayout.GetChild(i).gameObject);
+    
+                if (roomList[i].PlayerCount == 0)
+                {
+                    roomList.Remove(roomList[i]);
+                }
+            }
+        }
+
+
+        foreach (var room in roomList)
+        {
+            GameObject newRoom = Instantiate(roomNamePrefab, gridLayout.position, Quaternion.identity);
+
+            newRoom.GetComponentInChildren<Text>().text = room.Name + "(" + room.PlayerCount + ")";
+            newRoom.transform.SetParent(gridLayout);
+        }
     }
 }
