+++ ProjectSettings/EditorBuildSettings.asset
@@ -5,19 +5,19 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 0
+  - enabled: 1
     path: Assets/Scenes/Menu/MultiplayerMenu.unity
     guid: a26b5f0547247bf839d8e094c2ada06d
-  - enabled: 0
+  - enabled: 1
     path: Assets/Scenes/Menu/LobbyScene.unity
     guid: 9969625f187ed8483827e972b766c7d6
-  - enabled: 0
+  - enabled: 1
     path: Assets/Scenes/GameScene.unity
     guid: 9fc0d4010bbf28b4594072e72b8655ab
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/64/Login.unity
     guid: 9f90d213d559845bb8a139696f482150
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/64/Hall.unity
     guid: 13cff4a88c3f0414dbc07fc05c5d722b
   m_configObjects: {}
